Move quiz screen out of end screen
