+++ templates/vue/src/components/lightbox/quiz-screen/AnswerButton.vue
