+++ templates/vue/src/components/lightbox/quiz-screen/CompletionScreen.vue
