+++ templates/vue/src/components/lightbox/quiz-screen/Question.vue
